+++ expressEngine/lib/validationHandler/index.js
@@ -1,0 +1,5 @@
+import validateRequired from './validateRequired';
+
+export {
+	validateRequired,
+};
